Cleaning main.cpp, added mesh factory and did some changes that compiler suggested

+++ GameEngine/include/SystemUtils.h
@@ -5,7 +5,7 @@
 
 class SystemUtils {
 public:
-	static void GetDesktopResolution(float& horizontal, float& vertical) {
+	static void GetDesktopResolution(int& horizontal, int& vertical) {
 		RECT desktop;
 		const HWND hDesktop = GetDesktopWindow();
 		GetWindowRect(hDesktop, &desktop);
